refactor: 단일/다중 선택 상태를 훅으로 분리
